+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Release notes are generated from [Conventional Commits](https://www.conventionalcommits.org/)
 on the default branch using [git-cliff](https://git-cliff.org/).
 
+## [Unreleased]
+
+### Features
+
+- **scope:** Add cluster static-ref namespace allowlist [a8af5d4](https://github.com/konih/kollect/commit/a8af5d44c1ddb3ebfcddd29f52dcf8555e0162b1)
+
+
+### Refactoring
+
+- **controller:** Trim cluster target reconcile complexity [4d8e5ad](https://github.com/konih/kollect/commit/4d8e5ad9cac5967d09a8697fa249e5ce0525fcfe)
+
+- **controller:** Extract cluster scope enforcement [85daa4c](https://github.com/konih/kollect/commit/85daa4ccbc63ce983ee31ceba815fbb76c7df22d)
+
 ## [0.7.0-rc.1](https://github.com/konih/kollect/compare/v0.6.0-rc.2..v0.7.0-rc.1) - 2026-06-13
 
 ### Bug Fixes